Make it possible to add stagehand to new tables

+++ lib/stagehand/schema.rb
@@ -2,20 +2,29 @@ module Stagehand
   module Schema
     def self.add_stagehand!(options = {})
       ActiveRecord::Schema.define do
-        create_table :stagehand_commit_entries, :force => true do |t|
-          t.integer :record_id
-          t.string :table_name
-          t.string :operation, :null => false
-          t.integer :commit_id
-          t.string :session
+        unless Stagehand::Staging::CommitEntry.table_exists?
+          create_table :stagehand_commit_entries do |t|
+            t.integer :record_id
+            t.string :table_name
+            t.string :operation, :null => false
+            t.integer :commit_id
+            t.string :session
+          end
+
+          add_index :stagehand_commit_entries, :commit_id
+          add_index :stagehand_commit_entries, :operation
+
+          # Create trigger to initialize session using a function
+          ActiveRecord::Base.connection.execute("DROP TRIGGER IF EXISTS stagehand_session_trigger;")
+          ActiveRecord::Base.connection.execute("
+          CREATE TRIGGER stagehand_session_trigger BEFORE INSERT ON stagehand_commit_entries
+          FOR EACH ROW SET NEW.session = CONNECTION_ID();")
         end
 
-        add_index :stagehand_commit_entries, :commit_id
-        add_index :stagehand_commit_entries, :operation
-
         table_names = ActiveRecord::Base.connection.tables
-        table_names -= ['stagehand_commit_entries', 'schema_migrations']
+        table_names -= [Stagehand::Staging::CommitEntry.table_name, 'schema_migrations']
         table_names -= Array(options[:except]).collect(&:to_s)
+        table_names &= Array(options[:only]).collect(&:to_s) if options[:only].present?
 
         table_names.each do |table_name|
           Stagehand::Schema.drop_trigger(table_name, 'insert')
@@ -27,18 +36,12 @@ module Stagehand
           Stagehand::Schema.create_trigger(table_name, 'delete', 'OLD')
         end
       end
-
-      # Create trigger to initialize session using a function
-        ActiveRecord::Base.connection.execute("DROP TRIGGER IF EXISTS stagehand_session_trigger;")
-        ActiveRecord::Base.connection.execute("
-        CREATE TRIGGER stagehand_session_trigger BEFORE INSERT ON stagehand_commit_entries
-        FOR EACH ROW SET NEW.session = CONNECTION_ID();
-      ")
     end
 
-    def self.remove_stagehand!
+    def self.remove_stagehand!(options = {})
       ActiveRecord::Schema.define do
         table_names = ActiveRecord::Base.connection.tables
+        table_names &= Array(options[:only]).collect(&:to_s) if options[:only].present?
 
         table_names.each do |table_name|
           Stagehand::Schema.drop_trigger(table_name, 'insert')
@@ -46,7 +49,7 @@ module Stagehand
           Stagehand::Schema.drop_trigger(table_name, 'delete')
         end
 
-        drop_table :stagehand_commit_entries
+        drop_table :stagehand_commit_entries unless options[:only].present?
       end
     end
